// shared/calculators/interfaces/config_interfaces.ts
export interface OneEuroFilterConfig {
  frequency?: number; // Frequency of incoming frames defined in seconds. Used
  // only if can't be calculated from provided events (e.g.
  // on the very fist frame).
  minCutOff?: number; // Minimum cutoff frequency. Start by tuning this parameter while
  // keeping `beta=0` to reduce jittering to the desired level. 1Hz
  // (the default value) is a a good starting point.
  beta?: number; // Cutoff slope. After `minCutOff` is configured, start
  // increasing `beta` value to reduce the lag introduced by the
  // `minCutoff`. Find the desired balance between jittering and
  // lag.
  derivateCutOff?: number; // Cutoff frequency for derivate. It is set to 1Hz in the
  // original algorithm, but can be turned to further smooth the
  // speed (i.e. derivate) on the object.
  thresholdCutOff?: number; // The outlier threshold offset, will lead to a generally more
  // reactive filter that will be less likely to discount outliers.
  thresholdBeta?: number; // The outlier threshold slope, will lead to a filter
  // that will more aggressively react whenever the
  // keypoint speed increases by being less likely to
  // consider that an observation is an outlier.
  minAllowedObjectScale?: number; // If calculated object scale is less than given value smoothing
  // will be disabled and keypoints will be returned as is. This
  // value helps filter adjust to the distance to camera.
  disableValueScaling?: boolean; // Disable value scaling based on object size and use `1.0`
  // instead. Value scale is calculated as inverse value of object
  // size. Object size is calculated as maximum side of
  // rectangular bounding box of the object in XY plane.
}